Betty compliant

+++ 0x1A-sorting_algorithms/3-quick_sort.c
@@ -4,8 +4,6 @@
  * swap_int - function to swap the int at two memory addresses and prints array
  * @a: first int
  * @b: second int
- * @array: array to print
- * @size: size of array
  */
 void swap_int(int *a, int *b)
 {
@@ -17,12 +15,14 @@ void swap_int(int *a, int *b)
 /**
  * partition - splits an array into two arrays based on a pivot
  * @array: array to be partitioned
- * @size: size of the array
+ * @lo: starting index of the array
+ * @hi: last index of the array
+ * Return: index of
  */
 int partition(int *array, int lo, int hi)
 {
-	int p = array[hi];
-	int i = lo - 1, j = lo;
+	long int p = array[hi];
+	long int i = lo - 1, j = lo;
 	static int o_hi = 1;
 
 	if (o_hi < hi)
@@ -41,7 +41,12 @@ int partition(int *array, int lo, int hi)
 	print_array(array, o_hi);
 	return (i);
 }
-
+/**
+ * quicksort - helper function to split original function into more arguments
+ * @array: array to be sorted
+ * @lo: first index of array
+ * @hi: last index in array
+ */
 void quicksort(int *array, int lo, int hi)
 {
 	int p = 0;
